+++ noodles-util/src/variant/io.rs
@@ -4,8 +4,9 @@ mod compression_method;
 mod format;
 pub mod indexed_reader;
 pub mod reader;
+pub mod writer;
 
 pub use self::{
     compression_method::CompressionMethod, format::Format, indexed_reader::IndexedReader,
-    reader::Reader,
+    reader::Reader, writer::Writer,
 };