+++ noodles-util/src/variant/io/writer.rs
@@ -23,7 +23,7 @@ impl Writer {
     /// use noodles_vcf as vcf;
     /// use noodles_util::variant::{self, io::{CompressionMethod, Format}};
     ///
-    /// let mut writer = variant::writer::Builder::default()
+    /// let mut writer = variant::io::writer::Builder::default()
     ///     .set_format(Format::Bcf)
     ///     .set_compression_method(Some(CompressionMethod::Bgzf))
     ///     .build_from_writer(io::sink());
@@ -45,7 +45,7 @@ impl Writer {
     /// use noodles_vcf::{self as vcf, Record};
     /// use noodles_util::variant::{self, io::Format};
     ///
-    /// let mut writer = variant::writer::Builder::default()
+    /// let mut writer = variant::io::writer::Builder::default()
     ///     .set_format(Format::Vcf)
     ///     .set_compression_method(None)
     ///     .build_from_writer(io::sink());