+++ noodles-util/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+  * util/variant: Move readers (`Reader` and `IndexedReader`) to `io` module.
+
 ## 0.38.0 - 2024-03-12
 
 ### Changed